CRM v.2.1.0   * Added image thumbnails to conversations with clients and to the files tab in a contact profile.   * Added an always visible message date in conversations with clients.   * Added the display of clients’ company names in the phone call list.   * Made contact sorting by name more convenient.   * Improved delivery of web push notifications to mobile app users.   * Improved retrieving of incoming email messages and sending replies to them.   * Fixed citation folding in email messages.   * Improved import of country names in contact addresses from CSV files.   * Added a link to quickly open a Shop-Script order from a linked deal.   * Fixed sending of attached files in email notifications upon submission of web forms.   * Eliminated other found defects.

+++ api/v1/crm.file.delete.method.php
@@ -21,7 +21,22 @@ class crmFileDeleteMethod extends crmApiAbstractMethod
 
         $file_model->delete($file['id']);
 
-        $this->getMessageAttachmentsModel()->deleteByField(['file_id' => $file['id']]);
+        $message_link = $this->getMessageAttachmentsModel()->getByField(['file_id' => $file['id']]);
+        if (!empty($message_link)) {
+            $this->getMessageAttachmentsModel()->deleteByField(['file_id' => $file['id']]);
+            $message_footer_param = $this->getMessageParamsModel()->getByField([
+                'message_id' => $message_link['message_id'], 
+                'name' => 'footer',
+            ]);
+            $message_footer_param = empty($message_footer_param['value']) ? '' : $message_footer_param['value'] . '<br>';
+            $message_footer_param .= sprintf_wp('File <b>%s</b> was deleted by <b>%s</b>.', $file['name'], wa()->getUser()->getName());
+            $this->getMessageParamsModel()->replace([
+                'message_id' => $message_link['message_id'],
+                'name' => 'footer',
+                'value' => $message_footer_param,
+            ]);
+        }
+        
         $this->getNoteAttachmentsModel()->deleteByField(['file_id' => $file['id']]);
 
         $action = 'file_delete';
